requirements.txt
Flask==2.3.2
torch
torchvision
numpy==1.24.4
Pillow==9.5.0
python-dotenv==1.0.1
python-dotenv
requests
